Create dendrochronologial networks

Function to create dendrochronological networks based on the similarity. The input for this function is a similarity table created with the sim_table() function. The thresholds are set to the defaults (Visser 2021).

Usage

dendro_network(
  sim_table,
  r_threshold = 0.5,
  sgc_threshold = 0.7,
  p_threshold = 1e-04
)

Arguments

sim_table

a table of similarities created by the sim_table() function

r_threshold

all correlations equal or above this value are taken into account for the creation of edges. If you want all positive correlations included set this to 0.

sgc_threshold

all sgc-values equal or above this value are taken into account for the creation of edges.

p_threshold

all probabilities of exceedence equal or below this value are taken into account for the creation of edges.

Value

A simplified network of tree-ring material with the edges defined by the similarity.

References

Visser, RM. 2021a Dendrochronological Provenance Patterns. Network Analysis of Tree-Ring Material Reveals Spatial and Economic Relations of Roman Timber in the Continental North-Western Provinces. Journal of Computer Applications in Archaeology 4(1): 230–253. DOI: https://doi.org/10.5334/jcaa.79.

Examples

data(hol_rom)
sim_table_hol <- sim_table(hol_rom)
dendro_network(sim_table_hol)
#> IGRAPH 04bcf65 UN-- 13 10 -- 
#> + attr: name (v/c)
#> + edges from 04bcf65 (vertex names):
#>  [1] HOL_IRRBH --HOL_TMB11H  HOL_KBMBH --HOL_MBUFH   HOL_KORBH --HOL_MB253H 
#>  [4] HOL_LKBBH --HOL_MBRUH   HOL_LKBBH --HOL_MRBRH   HOL_MB253H--HOL_TRMB32H
#>  [7] HOL_MBRUH --HOL_MRBRH   HOL_MBUFH --HOL_MRBRH   HOL_MBUFH --HOL_PALZH  
#> [10] HOL_WEDEH --HOL_WEDWH  
dendro_network(sim_table_hol, r_threshold = 0.4, sgc_threshold = 0.6)
#> IGRAPH 1f1df32 UN-- 19 27 -- 
#> + attr: name (v/c)
#> + edges from 1f1df32 (vertex names):
#>  [1] HOL_AAHOH --HOL_MBUFH   HOL_IRRBH --HOL_TMB11H  HOL_KBMBH --HOL_MBUFH  
#>  [4] HOL_KBMBH --HOL_MRBRH   HOL_KBMBH --HOL_WEDEH   HOL_KBMBH --HOL_WEDWH  
#>  [7] HOL_KORBH --HOL_MB253H  HOL_KORBH --HOL_TATEH   HOL_KORBH --HOL_TRMB32H
#> [10] HOL_LKBBH --HOL_LKSBH   HOL_LKBBH --HOL_MBRUH   HOL_LKBBH --HOL_MRBRH  
#> [13] HOL_LKSBH --HOL_MRBRH   HOL_MB253H--HOL_TRMB32H HOL_MB253H--HOL_WEDEH  
#> [16] HOL_MB253H--HOL_WEDWH   HOL_MBRUH --HOL_MRBRH   HOL_MBRUH --HOL_TRRSH  
#> [19] HOL_MBUFH --HOL_MRBRH   HOL_MBUFH --HOL_PALZH   HOL_MBUFH --HOL_SQUSH  
#> [22] HOL_MBUFH --HOL_TRRSH   HOL_MRBRH --HOL_PALZH   HOL_PALZH --HOL_TRRSH  
#> + ... omitted several edges