Community detection using the Girvan-Newman algorithm

Function to determine the communities in a network using the Girvan-Newman algorithm. This function uses the cluster_edge_betweenness() function from the iGraph package, but creates a more user-friendly output that includes the names of the nodes.

Usage

gn_names(g)

Arguments

g

input graph or network that is used for community detection

Value

the names of the nodes in the various communities

Details

References Girvan, M and Newman, MEJ. 2002 Community structure in social and biological networks. Proceedings of the National Academy of Sciences of the United States of America 99(12): 7821–7826. DOI: https://doi.org/10.1073/pnas.122653799. Newman, MEJ and Girvan, M. 2004 Finding and evaluating community structure in networks. Physical Review E 69(2): 026113. DOI: https://doi.org/10.1103/PhysRevE.69.026113.

Examples

hol_sim <- sim_table(hol_rom)
g_hol <- dendro_network(hol_sim)
gn_names(g_hol)
#>           node com_name
#> 1    HOL_IRRBH     GN_1
#> 2   HOL_TMB11H     GN_1
#> 3    HOL_KBMBH     GN_2
#> 4    HOL_MBUFH     GN_2
#> 5    HOL_PALZH     GN_2
#> 6    HOL_KORBH     GN_3
#> 7   HOL_MB253H     GN_3
#> 8  HOL_TRMB32H     GN_3
#> 9    HOL_LKBBH     GN_4
#> 10   HOL_MBRUH     GN_4
#> 11   HOL_MRBRH     GN_4
#> 12   HOL_WEDEH     GN_5
#> 13   HOL_WEDWH     GN_5